+++ src/test/java/praktikum/BurgerMockTest.java
@@ -3,8 +3,10 @@ package praktikum;
 import org.junit.Test;
 import org.junit.runner.RunWith;
 import org.mockito.Mock;
-import org.mockito.Spy;
 import org.mockito.junit.MockitoJUnitRunner;
+
+import java.util.List;
+
 import static org.junit.Assert.*;
 
 import static org.mockito.Mockito.when;
@@ -17,7 +19,7 @@ public class BurgerMockTest {
     Ingredient ingredient;
     @Mock
     Ingredient ingredient1;
-    @Spy
+
     Burger burger = new Burger();
 
     @Test
@@ -31,6 +33,55 @@ public class BurgerMockTest {
         float expectedPrice = bun.getPrice() * 2 + ingredient.getPrice() + ingredient1.getPrice();//ожидаемый прайс бургера
         float actualPrice = burger.getPrice();//фактический прайс бургера
         assertEquals("Ожидаемая цена бургера не совпадает с фактической", expectedPrice, actualPrice, 0);
+    }
+
+    @Test
+    public void checkGetReceiptTest() {
+        burger.setBuns(bun);
+        when(bun.getName()).thenReturn("Black");
+        when(bun.getPrice()).thenReturn(100F);
+
+        burger.addIngredient(ingredient);
+        when(ingredient.getType()).thenReturn(IngredientType.SAUCE);
+        when(ingredient.getName()).thenReturn("BBQ");
+        when(ingredient.getPrice()).thenReturn(45F);
+
+        String expectedReceipt = String.format("(==== %s ====)%n", bun.getName()) + String.format("= %s %s =%n", ingredient.getType().toString().toLowerCase(),
+                ingredient.getName()) +
+                String.format("(==== %s ====)%n", bun.getName()) +
+                String.format("%nPrice: %f%n", burger.getPrice());
+        String actualReceipt = burger.getReceipt();
+        assertEquals("Ожидаемый чек не совпадает с фактическим", expectedReceipt, actualReceipt);
+    }
+
+    @Test
+    public void checkAddIngredientTest(){
+        burger.addIngredient(ingredient);
+        burger.addIngredient(ingredient1);
+        List<Ingredient> expectedIngredients = List.of(ingredient, ingredient1);
+        List<Ingredient> actualIngredients =  burger.ingredients;
+        assertEquals("Ожидаемые и фактические ингредиенты не совпадают",expectedIngredients, actualIngredients);
+    }
+
+    @Test
+    public void checkRemoveIngredientTest(){
+        burger.addIngredient(ingredient);
+        burger.addIngredient(ingredient1);
+        burger.removeIngredient(0);
+        List<Ingredient> expectedIngredients = List.of(ingredient1);
+        List<Ingredient> actualIngredients =  burger.ingredients;
+        assertEquals("Ожидаемые и фактические ингредиенты не совпадают",expectedIngredients, actualIngredients);
+        System.out.println(expectedIngredients);
+    }
+
+    @Test
+    public void checkMoveIngredientTest(){
+        burger.addIngredient(ingredient);
+        burger.addIngredient(ingredient1);
+        burger.moveIngredient(0, 1);
+        List<Ingredient> expectedIngredients = List.of(ingredient1, ingredient);
+        List<Ingredient> actualIngredients =  burger.ingredients;
+        assertEquals("Ожидаемые и фактические ингредиенты не совпадают",expectedIngredients, actualIngredients);
     }
 
 }
